requirements.txt
anyio==3.7.0
asyncpg==0.27.0
certifi==2023.5.7
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
exceptiongroup==1.1.1
fastapi==0.98.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
pycparser==2.21
pydantic==1.10.9
PyJWT==2.1.0
pytest==7.4.0
pytest-asyncio==0.21.0
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.6.3
uvicorn==0.22.0
